Add physipkpd to an existing project!
Can now add PhysiPKPD to an existing project. A lot of work went into finding the right place to add to the Makefile, main.cpp, and custom.h. Also, dealing with calling the script from the project directory itself.

+++ add_physipkpd.py
@@ -60,41 +60,56 @@ temp_dir = append_suffix("PhysiPKPD-TEMP")
 with ZipFile(local_file, 'r') as zObject: 
     zObject.extractall(path=temp_dir)
 
+os.remove(local_file)
+
 print(f"extracted PhysiPKPD to {temp_dir}")
 
 folder_name = os.listdir(f"./{temp_dir}")[0]
 
 import shutil
 
-os.rename(f"{temp_dir}/{folder_name}/addons/PhysiPKPD",f"{DIR}/addons/PhysiPKPD")
-os.removedirs(f"{temp_dir}/{folder_name}/addons")
-# os.rename(f"{temp_dir}/{folder_name}/sample_projects_physipkpd",f"{DIR}/sample_projects_physipkpd") # don't need to move sample projects in this case
-os.rename(f"{temp_dir}/{folder_name}/LICENSE",f"{DIR}/addons/PhysiPKPD/LICENSE")
-os.rename(f"{temp_dir}/{folder_name}/README.md",f"{DIR}/addons/PhysiPKPD/README.md")
+if os.path.exists(f"{DIR}/addons/PhysiPKPD") is False:
+    os.rename(f"{temp_dir}/{folder_name}/addons/PhysiPKPD",f"{DIR}/addons/PhysiPKPD")
+    os.removedirs(f"{temp_dir}/{folder_name}/addons")
+    # os.rename(f"{temp_dir}/{folder_name}/sample_projects_physipkpd",f"{DIR}/sample_projects_physipkpd") # don't need to move sample projects in this case
+    os.rename(f"{temp_dir}/{folder_name}/LICENSE",f"{DIR}/addons/PhysiPKPD/LICENSE")
+    os.rename(f"{temp_dir}/{folder_name}/README.md",f"{DIR}/addons/PhysiPKPD/README.md")
+    print(f"Moved PhysiPKPD files to {DIR}. Deleted {temp_dir}")
+else:
+    print(f"Found PhysiPKPD files already in {DIR}.")
 shutil.rmtree(temp_dir)
-print(f"Moved PhysiPKPD files to {DIR}. Deleted {temp_dir}")
+
+makefile = f'{DIR}/Makefile'
+with open(makefile, 'r+') as f:
+    lines = f.readlines()
+    for i, line in enumerate(lines):
+        if "PhysiCell_custom_module_OBJECTS :=" in line:
+            lines.insert(i,"PhysiPKPD_OBJECTS := PhysiPKPD_PK.o PhysiPKPD_PD.o\n\n")
+            f.seek(0)
+            f.writelines(lines)
+            break
 
 source_file = f'{DIR}/addons/PhysiPKPD/Makefile-PhysiPKPD-Objects.txt'
 with open(source_file, 'r') as sf:
     new_lines = sf.readlines()
-print(new_lines)
-exit()
-with open(f'{DIR}/Makefile', 'r') as f:
+
+with open(makefile, 'r') as f:
     lines = f.readlines()
 
 for line_number, line in enumerate(lines):
-    if "PhysiCell_custom_module_OBJECTS :=" in line:
+    if "custom_modules/custom.cpp" in line:
         break
 
-with open(f'{DIR}/Makefile','w') as f:
+with open(makefile,'w') as f:
     for i in range(line_number):
-        print(lines[i], file=f)
+        print(lines[i].rstrip('\n'), file=f)
     for new_line in new_lines:
-        print(new_line, file=f)
+        print(new_line.rstrip('\n'), file=f)
+    print("\n", file=f)
     for i in range(i+1,len(lines)):
-        print(lines[i], file=f)
+        print(lines[i].rstrip('\n'), file=f)
 
-with open(f'{DIR}/Makefile','r+') as f:
+with open(makefile,'r+') as f:
     lines = f.readlines()
 
 for i, line in enumerate(lines):
@@ -103,7 +118,7 @@ for i, line in enumerate(lines):
         break
 lines[i] = line
 
-with open(f'{DIR}/Makefile','w') as f:
+with open(makefile,'w') as f:
     f.writelines(lines)
 
 print(f"Updated Makefile to be ready for PhysiPKPD")
@@ -177,34 +192,35 @@ if args.studio:
     with open(local_file, 'wb')as file:
         file.write(data.content)
 
-    studio_dir = append_suffix(f"{DIR}-studio")
-    studio_dir_temp = append_suffix(f"{DIR}-studio-TEMP")
+    dir_name = os.path.abspath(DIR).split(os.sep)[-1] # protects aginst using the current working directory "."
+    studio_dir = append_suffix(f"{DIR}/../{dir_name}-studio")
+    studio_dir_temp = append_suffix(f"{studio_dir}-TEMP")
     with ZipFile(local_file, 'r') as zObject: 
         zObject.extractall(path=studio_dir_temp)
+    os.remove(local_file)
     folder_name = os.listdir(f"./{studio_dir_temp}")[0]
     os.rename(f"{studio_dir_temp}/" + folder_name, studio_dir)
 
     os.removedirs(studio_dir_temp)
 
+    with open(makefile, "r") as f:
+        for line in f:
+            if "PROGRAM_NAME :=" in line:
+                project_name = line.split(" ")[-1].strip('\n')
+
 print(f"You are all set!")
 print(f"\t1. Move into your project folder:")
 print(f"\t\tcd {DIR}")
-print(f"\t2. Make a sample project or a template project and make your PhysiPKPD model! Examples:")
-print(f"\t\tmake pkpd-proliferation-sample\n\t\tmake pkpd-apoptosis-sample\n\t\tmake pkpd-template\n\n")
-print(f"\t3. Compile the project:")
+print(f"\t2. (Re-)compile the project:")
 print(f"\t\tmake -j 8")
 if args.studio:
-    print(f"\t4. Edit them with studio:")
+    print(f"\t3. Edit them with studio. Make sure your config file is correct:")
     print(f"\t  For a sample project:")
-    print(f"\t\t(MacOS/Unix) python ../{studio_dir}/bin/studio.py -c ./config/pkpd_model.xml -e pkpd_sample --pkpd")
-    print(f"\t\t(Windows) python ..\{studio_dir}\bin\studio.py -c .\config\pkpd_model.xml -e pkpd_sample.exe --pkpd\n")
+    print(f"\t\t(MacOS/Unix) python {studio_dir}/bin/studio.py -c ./config/PhysiCell_settings.xml -e {project_name} --pkpd")
+    print(f"\t\t(Windows) python {studio_dir}\bin\studio.py -c .\config\PhysiCell_settings.xml -e {project_name}.exe --pkpd\n")
     print(f"\t  For a template project:")
-    print(f"\t\t(MacOS/Unix) python ../{studio_dir}/bin/studio.py -c ./config/pkpd_model.xml -e pkpd_project --pkpd")
-    print(f"\t\t(Windows) python ..\{studio_dir}\bin\studio.py -c .\config\pkpd_model.xml -e pkpd_project.exe --pkpd\n\n")
+    print(f"\t\t(MacOS/Unix) python {studio_dir}/bin/studio.py -c ./config/PhysiCell_settings.xml -e {project_name} --pkpd")
+    print(f"\t\t(Windows) python {studio_dir}\bin\studio.py -c .\config\PhysiCell_settings.xml -e {project_name}.exe --pkpd\n\n")
 else:
-    print(f"\t4a. Run the samples with")
-    print("\t\t(MaxOS/Unix) ./pkpd_sample ./config/pkpd_model.xml")
-    print(f"\t\t(Windows) pkpd_sample.exe .\config\pkpd_model.xml")
-    print(f"\t4b. Run the template projects with")
-    print(f"\t\t (MaxOS/Unix) ./pkpd_project ./config/pkpd_model.xml")
-    print(f"\t\t (Windows) pkpd_project.exe .\config\pkpd_model.xml")
+    print(f"\t3. Edit according to https://github.com/drbergman/PhysiPKPD/releases/latest")
+    print(f"Consider passing in the --studio flag next time to get studio set up for easier editing")
